requirements.txt
todoist-api-python==2.1.7
openai==1.4.0
pyyaml
coverage==7.3.2
pytest==7.4.3
pytest-mock==3.14.0
python-dotenv==1.0.0